Added transactions feature.
Only works inside one quorum by locking the major key on the primary. This is not replicated, so primary failover effectively rolls back pending transactions. It is guaranteed that SETs and DELETEs of a transaction are replicated in one round. The client library only sends writes when the CommitTransaction() is issued. All other commands bypass the transactions feature but reads see issued writes, as it works with the proxy batching feature.

Added StartTransaction, CommitTransaction and RollbackTransaction to client libraries (C#, Java, Python).